requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.1
pytest==9.0.2
python-dateutil==2.9.0.post0
scipy==1.17.0
six==1.17.0
